Tizen 2.4.0 rev3 SDK Public Release

+++ include/mobile/player_internal.h
@@ -17,6 +17,7 @@
 #ifndef __TIZEN_MEDIA_PLAYER_INTERNAL_H__
 #define	__TIZEN_MEDIA_PLAYER_INTERNAL_H__
 #include <player.h>
+#include <mm_player.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -42,7 +43,7 @@ typedef struct
 	int rate;            /**<  Samplerate */
 	int depth;           /**< Depth */
 	bool little_endian;  /**< Endianness */
-	unsigned long long channel_mask;	  /**< channel_mask */
+	guint64 channel_mask;	  /**< channel_mask */
 } player_audio_raw_data_s;
 
 /**
